pyproject.toml
[tool.poetry]
name = "Bytovy Lovec"
version = "1.0.0"
description = "All pipelines and modules for Bytovy Lovec"
authors = ["Adam Bako <adam@radiobotics.com>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"

####################
# SHARED LIBRARIES #
####################
ai_client = { path = "shared/ai_client" }
base_ad_scraper = { path = "shared/base_ad_scraper" }
bezrealitky_scraper = { path = "shared/bezrealitky_scraper" }
postgres_client = { path = "shared/postgres_client" }
reality_idnes_scraper = { path = "shared/reality_idnes_scraper" }
rental_ad_schema = { path = "shared/rental_ad_schema" }
sreality_scraper = { path = "shared/sreality_scraper" }
dbt_runner = { path = "shared/dbt_runner" }
email_bot = { path = "shared/email_bot" }

##############
# LIBRARIES  #
##############
selenium = "~4.27.1"
pydantic = "~2.10.5"
setuptools = "~70.0.0"
psycopg2-binary = "~2.9.1"
flask = "~3.1.0"
google-generativeai = "0.8.4"
prefect_docker = "0.6.2"
dbt-core = "1.8.1"
protobuf = "4.25.3"
tabulate = "0.9.0"
dbt-postgres = "1.9.0"
resend = "2.6.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"